Move function to a contributor module.
