+++ elifecrossref/generate.py
@@ -10,7 +10,7 @@ from elifearticle import parse
 from elifetools import utils as etoolsutils
 from elifetools import xmlio
 
-from elifecrossref import utils, elife
+from elifecrossref import utils, elife, contributor
 from elifecrossref.conf import raw_config, parse_raw_config
 from elifecrossref.mime_type import crossref_mime_type
 from elifecrossref.tags import add_clean_tag, add_inline_tag, clean_tags
@@ -149,8 +149,8 @@ class CrossrefXML(object):
         # Set the title with italic tag support
         self.set_titles(journal_article_tag, poa_article)
 
-        set_contributors(journal_article_tag, poa_article,
-                         self.crossref_config.get("contrib_types"))
+        contributor.set_contributors(journal_article_tag, poa_article,
+                                     self.crossref_config.get("contrib_types"))
 
         self.set_abstract(journal_article_tag, poa_article)
         self.set_digest(journal_article_tag, poa_article)
@@ -691,70 +691,6 @@ def has_license(poa_article):
     if not poa_article.license.href:
         return False
     return True
-
-
-def set_contributors(parent, poa_article, contrib_types=None):
-    # First check for any contributors
-    if not poa_article.contributors:
-        return
-    # If contrib_type is None, all contributors will be added regardless of their type
-    contributors_tag = SubElement(parent, "contributors")
-
-    # Ready to add to XML
-    # Use the natural list order of contributors when setting the first author
-    sequence = "first"
-    for contributor in poa_article.contributors:
-        if contrib_types:
-            # Filter by contrib_type if supplied
-            if contributor.contrib_type not in contrib_types:
-                continue
-
-        if contributor.contrib_type == "on-behalf-of":
-            contributor_role = "author"
-        else:
-            contributor_role = contributor.contrib_type
-
-        # Skip contributors with no surname
-        if contributor.surname == "" or contributor.surname is None:
-            # Most likely a group author
-            if contributor.collab:
-                organization_tag = SubElement(contributors_tag, "organization")
-                organization_tag.text = contributor.collab
-                organization_tag.set("contributor_role", contributor_role)
-                organization_tag.set("sequence", sequence)
-
-        else:
-            person_name_tag = SubElement(contributors_tag, "person_name")
-
-            person_name_tag.set("contributor_role", contributor_role)
-
-            person_name_tag.set("sequence", sequence)
-
-            given_name_tag = SubElement(person_name_tag, "given_name")
-            given_name_tag.text = contributor.given_name
-
-            surname_tag = SubElement(person_name_tag, "surname")
-            surname_tag.text = contributor.surname
-
-            if contributor.suffix:
-                suffix_tag = SubElement(person_name_tag, "suffix")
-                suffix_tag.text = contributor.suffix
-
-            if contributor.affiliations:
-                # Crossref schema limits the number of affilations an author can have
-                max_affiliations = 5
-                for aff in contributor.affiliations[0:max_affiliations]:
-                    if aff.text and aff.text != '':
-                        affiliation_tag = SubElement(person_name_tag, "affiliation")
-                        affiliation_tag.text = aff.text
-
-            if contributor.orcid:
-                orcid_tag = SubElement(person_name_tag, "ORCID")
-                orcid_tag.set("authenticated", "true")
-                orcid_tag.text = contributor.orcid
-
-        # Reset sequence value after the first sucessful loop
-        sequence = "additional"
 
 
 def set_publication_date(parent, pub_date):
